ported to MS-DOS (kind of)
The low system requirements and high platform-agnosticism in my design of this software have tempted me to see how it compiles for a 16-bit DOS system using the Watcom C compiler.

Trivial warnings (like not being able to shift (int)(0) by more than 15 bits) have all been fixed.

The only detail left unported is that the Nintendo 64 flash memory chip is addressable up to a maximum of byte address 0x01FFFF, while real-mode DOS applications run on a completely 16-bit environment with a maximum memory address of 0x00FFFF, which leaves a big problem with having a fully working MS-DOS port of the program.  At least now it will work with save files half of the maximum possible size Project64 might export and import.

I don't really care to fully port it to MS-DOS to the point of splitting the file into two flash buffers or anything that extreme.  The trial of portability has been run completely.

+++ flash_io.c
@@ -182,7 +182,7 @@ unsigned int swap_flash(unsigned int interval)
 {
     RCP_block RCP;
     unsigned int mask;
-    register size_t i, j;
+    register u32 i, j;
 
     if (interval != 0) { /* callee-defined endian swap on a fixed interval */
         mask = interval - 1; /* e.g. (interval = 4) for a 32-bit swap */
